@@ -8,6 +8,26 @@ using FluentAssertions;
 namespace Othello.Tests {
 	public class Tests {
 		[Fact]
+		public void NewBoard() {
+			// Test setup of new board.
+			OthelloBoard board = new OthelloBoard();
+			foreach (var pos in BoardPosition.GetRectangularPositions(8, 8)) {
+				// (3, 3) and (4, 4) are white (player 2).
+				if (pos is { Row: 3, Col: 3 } or { Row: 4, Col: 4 }) {
+					board.GetPlayerAtPosition(pos).Should().Be(2);
+				}
+				// (3, 4) and (4, 3) are black (player 1).
+				else if (pos is { Row: 3, Col: 4 } or { Row: 4, Col: 3 }) {
+					board.GetPlayerAtPosition(pos).Should().Be(1);
+				}
+				// All others are empty.
+				else {
+					board.GetPlayerAtPosition(pos).Should().Be(0);
+				}
+			}
+		}
+
+		[Fact]
 		public void NewValue() {
 			// New board Value is 0.
 			OthelloBoard board = new OthelloBoard();
@@ -22,14 +42,24 @@ namespace Othello.Tests {
 		}
 
 		[Fact]
-		public void ValueAfterMultiDirectionMove() {
-			// Value updated correctly after doing a multi-directional move.
+		public void MultiDirectionMove() {
+			// Value and board updated correctly after doing a multi-directional move.
 			OthelloBoard board = new OthelloBoard();
-			board.ApplyMove(new OthelloMove(new BoardPosition(3, 2)));
-			board.ApplyMove(new OthelloMove(new BoardPosition(4, 2)));
+			board.ApplyMove(new BoardPosition(3, 2));
+			board.ApplyMove(new BoardPosition(4, 2));
 			board.CurrentAdvantage.Advantage.Should().Be(0);
-			board.ApplyMove(new OthelloMove(new BoardPosition(5, 2)));
+			board.ApplyMove(new BoardPosition(5, 2));
 			board.CurrentAdvantage.Advantage.Should().Be(5);
+			
+			// Check both "up" and "up-right" directions to make sure pieces were flipped.
+			board.GetPlayerAtPosition(new BoardPosition(5, 2)).Should().Be(1);
+			board.GetPlayerAtPosition(new BoardPosition(4, 2)).Should().Be(1);
+			board.GetPlayerAtPosition(new BoardPosition(3, 2)).Should().Be(1);
+			board.GetPlayerAtPosition(new BoardPosition(4, 3)).Should().Be(1);
+			board.GetPlayerAtPosition(new BoardPosition(3, 4)).Should().Be(1);
+
+			// But not the one remaining black piece.
+			board.GetPlayerAtPosition(new BoardPosition(4, 4)).Should().Be(2);
 		}
 	}
 }
